test/backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
openai>=1.0
python-dotenv>=1.0
